Use float16 as routing logits dtype
Otherwise, the benchmark is needlessly slow owing to the use of
float32 logits. This change improves the runtime from 22.8us to
17.3us.

+++ python/triton_kernels/tests/test_routing.py
@@ -5,7 +5,7 @@ from triton_kernels.testing import assert_close
 from triton_kernels.testing import assert_equal
 
 
-def init_data(n_tokens, n_expts_tot, dtype=torch.float32, device="cuda"):
+def init_data(n_tokens, n_expts_tot, dtype=torch.float16, device="cuda"):
     logits = torch.randn((n_tokens, n_expts_tot), dtype=dtype, device=device, requires_grad=True)
     return logits
